[订餐系统] 整体流程已经跑通

+++ pojo/src/main/java/com/yinrj/dto/OrderMessageDTO.java
@@ -56,7 +56,7 @@ public class OrderMessageDTO {
     /**
      * 积分数量
      */
-    private Integer rewardAmount;
+    private BigDecimal rewardAmount;
 
     /**
      * 确认
